revert lifecycle changes to server
